requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.0.350
langchain-openai>=0.0.5
langchain-core>=0.1.0
python-dotenv>=1.0.0
openai>=1.3.0

# HTTP requests
requests>=2.31.0

# Data handling
pydantic>=2.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
